Layer layout: change how backward links are displayed.
Previously, links were attached to an entry's side depending on whether the other
entry was in a lower or higher layer.

With this commit, links going out of an entry are all attached to outbound slots,
while links going to the entry are attached to inbound slots. Said another way,
all edges & vertex inputs are on the same side, and all outputs are on the other
side.